+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Funcionalidade: Adicionada exibição de reservas de acordo com o auditório selecionado.
 * Funcionalidade: Adicionada função de abrir modal de reserva com horário de início preenchido.
 * Funcionalidade: Adicionada confirmação sobre termos de Local para criação de Reserva.
+* Funcionalidade: Adicionada opção de upload de imagem para local.
 
 # 0.0.2
 * Início versionamento.